# Improve Value and PathElement String()