@lavalamp 2019-10-02
> PathElements are frequently turned into strings to find out if a path has already been evaluated

Where? This shouldn't be done?