@lavalamp 2019-10-02
We should just store the path or path element and use Less (and/or use or implement Equals if not already).